switch to floats and add timing

+++ sequential/color.h
@@ -6,8 +6,8 @@
 
 using color = vec3;
 
-inline double linear_to_gamma(double linear_component) {
-  return sqrt(linear_component);
+inline float linear_to_gamma(float linear_component) {
+  return sqrtf(linear_component);
 }
 
 void write_color(std::ostream &out, color pixel_color, int samples_per_pixel) {
@@ -16,7 +16,7 @@ void write_color(std::ostream &out, color pixel_color, int samples_per_pixel) {
     auto b = pixel_color.z();
 
     // Divide the color by the number of samples.
-    auto scale = 1.0 / samples_per_pixel;
+    auto scale = 1.0f / samples_per_pixel;
     r *= scale;
     g *= scale;
     b *= scale;
@@ -28,10 +28,10 @@ void write_color(std::ostream &out, color pixel_color, int samples_per_pixel) {
 
 
     // Write translated [0, 255] value of each color component
-    static const interval intensity(0.000, 0.999);
-    out << static_cast<int>(255.999 * intensity.clamp(r)) << ' '
-        << static_cast<int>(255.999 * intensity.clamp(g)) << ' '
-        << static_cast<int>(255.999 * intensity.clamp(b))<< '\n';
+    static const interval intensity(0.000f, 0.999f);
+    out << static_cast<int>(255.999f * intensity.clamp(r)) << ' '
+        << static_cast<int>(255.999f * intensity.clamp(g)) << ' '
+        << static_cast<int>(255.999f * intensity.clamp(b))<< '\n';
 }
 
 #endif
