+++ sequential/hittable.h
@@ -8,7 +8,7 @@ class hit_record {
 public:
     point3 p;
     vec3 normal;
-    double t;
+    float t;
     bool front_face;
 
     void set_face_normal(const ray& r, const vec3& outward_normal) {
